refactor: organize files and simplifying some basic validations Co-authored-by-josueHernandez
