+++ src/Smartflow.Domain/Enums/AlertSeverity.cs
@@ -1,0 +1,9 @@
+namespace Smartflow.Domain.Enums
+{
+  public enum AlertSeverity
+  {
+    INFO,
+    WARNING,
+    CRITICAL
+  }
+}
