+++ src/Smartflow.Domain/Enums/ParallelizationStrategy.cs
@@ -1,0 +1,8 @@
+namespace Smartflow.Domain.Enums
+{
+  public enum ParallelizationStrategy
+  {
+    DATA_DECOMPOSITION,
+    TASK_PARALLELISM
+  }
+}
