+++ src/Smartflow.Domain/Enums/SensorType.cs
@@ -1,0 +1,11 @@
+namespace Smartflow.Domain.Enums
+{
+  public enum SensorType
+  {
+    NOISE,
+    TEMP,
+    HUMIDITY,
+    TRAFFIC,
+    POLLUTION
+  }
+}
